examples/st-oid4vci/demo/requirements.txt
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
idna==3.7
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
setuptools==70.3.0
urllib3==2.2.2
wheel==0.43.0
